Refactored multi func method

+++ TopSolutions.ConsoleApp/GeneralCSharp/Func/Calculator.cs
@@ -9,9 +9,9 @@ namespace TopSolutions.ConsoleApp.GeneralCSharp.Func
     public class Calculator
     {
         public double Add(int a, int b) => a + b;
-        public int Subtract(int a, int b) => a - b;
+        public double Subtract(int a, int b) => a - b;
         public double Multiply(int a, int b) => a * b;
-        public int Divide(int a, int b) => a / b;
+        public double Divide(int a, int b) => a / b;
         public double PowerOf(int a, int b) => Math.Pow(a, b);
     }
 }
