+++ TopSolutions.ConsoleApp/GeneralCSharp/Func/FuncMain.cs
@@ -13,9 +13,9 @@ namespace TopSolutions.ConsoleApp.GeneralCSharp.Func
             var funcCalculator = new Calculator();
 
             Func<int, int, double> add = funcCalculator.Add;
-            Func<int, int, int> subtract = funcCalculator.Subtract;
+            Func<int, int, double> subtract = funcCalculator.Subtract;
             Func<int, int, double> multiply = funcCalculator.Multiply;
-            Func<int, int, int> divide = funcCalculator.Divide;
+            Func<int, int, double> divide = funcCalculator.Divide;
             Func<int, int, double> powerOf = funcCalculator.PowerOf;
             Console.WriteLine($"Addition result: {add(410000, 255220)}");
             Console.WriteLine($"Subtraction result: {subtract(4, 2)}");
